Refactor dev dependency handling (#701)
* Switch dev deps from an extra to pip-compile/multi

All dev dependencies are herein moved out of the `globus-sdk[dev]`
extra, and into files in a `requirements/` dir.
These files are managed using pip-compile-multi and pip-compile,
which can be run near-trivially with a new tox env:
`tox r -e freezedeps`.

The pip-compile-multi workflow has been put forth in the pallets
projects where it is well liked for providing build
reproducibility, and is used by several other community projects
as well. However, for our "mindeps" environments, pip-compile-multi is
unsuitable because it explicitly protects against mismatched
dependency versions.

Why do this rearrangement? And why now?

To the first question, this removes our development needs from
published package metadata (extras). This is good for long-term
maintenance: it simplifies what we are publishing to only what is
useful, necessary, and supported, and the change means that the
various collections of dev dependencies do not need to be installable
into any singular virtualenv (e.g. 'docs' needs can conflict with
'test').

To the second question, an issue has arisen with the introduction of
sphinx-inlne-tabs for the doc build. It currently requires python3.8+,
but we are still testing on python3.7 . This needs some resolution,
and this one is best because it provides us other benefits.

Some additional details shake loose here:
- unnecessary cast due to 'responses' update (mypy)
- The requests version can't be kept at 2.19.1 by this pinning
  technique for the mindeps build because that would also pin back the
  responses version and break tests. For now this is handled by using
  a slightly newer version, which might be for the best in any case

* Update with changes from review

Co-authored-by: Kurt McKee <contactme@kurtmckee.org>

* Replace pip-compile-multi with pip-compile script

We are working against pip-compile-multi almost as much as we're
working with it. Rather than continue, solve the desire for
CUSTOM_COMPILE_COMMAND and resolve the awkward split between
requirements by introducing a very small wrapper which uses
pip-compile in the way which we'd like.

* Relativize frozen deps within requirements/

---------

Co-authored-by: Kurt McKee <contactme@kurtmckee.org>