Begin using tabs to refactor method docs (#700)
* Begin using tabs to refactor method docs

This is a combined set of several minor changes which support a new
way of writing docs for client methods. The overall goal is to make
use of sphinx-inline-tabs to present several pieces of information in
a way which is compact by default, expandable, and allows us to put
some more structure to what is currently a long scroll of text.

To wit:
- include the sphinx-inline-tabs extension for sphinx
- cleanup our custom sphinx directives to use docutils' built-in
  option parsing
- add new custom directives for several small functions:
  - expandtestfixture: expands a JSON or text body from a test fixture
    as a code-block, for use as example data
  - extdoclink: link to an external doc location (only supports
    docs.globus.org right now)
  - paginatedusage: emit the standard paginated usage info text
- select two methods which are updated to use the new documentation
  capabilities: TransferClient.get_submission_id and
  SearchClient.search. These demonstrate the behavior and what is now
  possible

It should be noted that `get_submission_id` has no usage example while
`search` does, that `search` is paginated while `get_submission_id` is
not, and that both refer to docs site docs and have relevant test
fixture data for examples. Therefore, they demonstrate the flexibility
of this setup to handle different kinds of methods, but also share
some commonalities.

The original drive of this work was `expandtestfixture`, but once that
was achieved, it demands some solution for better presentation, like
tabs. Once tabs are in place, they conflict with the default behavior
for the pagination decorator (which appends to the docstring), so some
accommodation is needed. A similar need precipitates the refactor of
the external doc decorator into a sphinx directive.

* Apply suggestions from code review

Co-authored-by: Kurt McKee <contactme@kurtmckee.org>

* Put a small border around tab groups in docs

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* mypy fix: guard against __doc__ being None

* Rename API Info section in doc tabs + add padding

With the API route info now inside of the tab, the left border clips
with the code styling and reveals a problem with the spacing on the
borders here.

To resolve, pad the content of the tab cells (not the whole tab
container).

* Switch to sphinx-design; tabify all paginatedusage

blacken-docs was breaking on the updated pagination doc modifying
decorator. Part of the point of this overall changeset is to remove
that decorator altogether. Therefore, after a quick assessment (under
20 methods to touch), remove that feature from the paginator and apply
the inline-tab usage across all paginated methods, putting the
relevant API Info, Paginated Usage, and Example Usage sections into
tabs.

This is a significant rewrite of several method docs but allows us to
remove a "second" path for paginated method docs.

Several discoveries were driven by this:
- sphinx-inline-tabs synchronizes all tabs with the same name, which
  is highly disruptive for our docs, and has no way to turn this
  behavior off. Switch to sphinx-design for our tab implementation,
  which fixes this issue and is cosmetically identical under furo
- Many of our examples were not targeted by blacken-docs and several
  syntax and usage fixed for our pycon blocks allows blacken-docs to
  apply, rewriting many examples
- Many of these examples, once blackened, are too long for our length
  limit. Cover the gaps with noqa:E501 comments

* Apply suggestions from code review

Co-authored-by: Kurt McKee <contactme@kurtmckee.org>

---------

Co-authored-by: Kurt McKee <contactme@kurtmckee.org>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>