Refactor error parsing to better support JSON:API
Core error parsing has been rewritten to be a better match for the
multiple-errors paradigm of JSON:API and similar APIs.

Parsing is now broken down internally into the following phases:
- _parse_response: the main parsing method, which aborts or calls the
  following methods, listed in order
- _parse_errors_array: extract one or more sub-errors from the
  document
- _parse_code: extract a singular "code" field from the document
- _parse_messages: extract message strings from the document

GlobusAPIError and its subclasses now feature an `errors` attribute,
which is a list of sub-errors, and a `messages` attribute, which is a
list of message strings. `message` is a property which is defined as
`"; ".join(messages)`, and has a deprecated setter.

`message` will be `None` when there is no parsed message, as
opposed to the previous behavior which falls back to the text of the
response.

`code` parsing can consequently be improved to have clearer semantics
with respect to multiple sub-errors with different codes -- in this
case, it will no longer assume that the first code from the sub-errors
is accurate. This means that we won't show misleading results if an
errors array comes back like
    {
      "errors": [
        {"code": "foo"},
        {"code": "bar"}
      ]
    }

Previously, this would parse the code as `foo`, but now it will leave
it with the default of `Error`.