Allow 'required_scope' singular variant (#726)
* Allow 'required_scope' singular variant

ConsentRequiredInfo now accepts 'required_scope' containing an
individual string as an alternative to 'required_scopes'. The
validation of these values has also been made slightly stricter.

Additionally, `ConsentRequiredInfo.required_scopes` is now the empty
list in the falsy case. This is not a very significant change, since

1. the documented usage pattern is that the truthiness of
  `ConsentRequiredInfo` should always be checked before it is used
2. the resulting value is still itself falsy

This subtlety is not covered in the changelog, on the grounds that it
would only bring undue attention to a minor implementation detail.

* Change test error response code to match spec