Allow list_flows to take orderby as a tuple
The new type for orderby allows str or
`tuple[str, Literal["ASC", "DESC"]]`.

The result is that the new ordering style can be validated by mypy,
but the existing string style is kept in place.

The validation is intentionally very limited, and primarily provides a
nicety for any type-based editor integrations and type checked
codebases. The goal is not to fully validate possible `orderby` values.