Make the treatment of JSON:API errors stricter
Detect and parse JSON:API errors based on the mimetype as specified in
the draft Globus Error Specification.

This fully separates the three parsing paths for JSON:API, Type Zero,
and Undefined.

Flows and Timers errors now customize their Undefined parsing methods.

In accordance with the draft spec, request_id is now an attribute of
all parsed error objects.

The mimetype detection revealed a bug in `globus_sdk._testing` in
which the Content-Type header was being set twice for JSON data. As a
result, mimetype parsing did not behave as desired and expected.