For non-JSON:API errors, parse an 'errors' array
For both the Type Zero and Undefined parse paths, 'errors' is now
defined as the root document if there is no 'errors' key or if the
contents of the 'errors' key are not 100% valid.

This tightens up the logic in the undefined path and adds new
capabilities to the Type Zero path.