Implement and document ErrorSubdocument for errors
GlobusAPIError.errors is now a list of ErrorSubdocument objects. These
are documented with an ultra-simple usage example as part of the core
exception documentation.

ErrorSubdocuments are defined to match well to the JSON:API notion of
an error object, but reduced to the fields which are appropriate to
our usage: code and message.

Because the `errors` objects now have some capabilities, some of the
parsing around (e.g., `messages` extraction) can be simplified.
`ErrorSubdocument` objects contain the list of fields they use for
`message` lookup, and therefore need to be customized to use
`GlobusAPIGlobusAPIError.MESSAGE_FIELDS` in a few locations.

Test changes are minimal in order to align types -- no test
behaviors are adjusted here other than those to fix the type of
`errors`.