Update requirements (#733)
* Run `piptools compile` in the `requirements/` dir

This aligns with how dependabot generates updates.

* Upgrade dependencies when freezing them

* Upgrade the requirements files

* Fix test with bad Content-Length (#735)

Under urllib3 v2.0, a Content-Length longer than the response content
(leading to an incomplete read) is no longer silently ignored, but
raises an error. This surfaced in an apparently unrelated test in
which the body was 131 bytes, but Content-Length was specified at 138.

We could correct the Content-Length, but it is not essential to the
test at all. Remove it.

---------

Co-authored-by: Stephen Rosen <sirosen@globus.org>