Update authorization params parsing: support list[str] for policies and more type safety (#769)
* Allow session_required_policies to be list[str]

When parsing authorization_parameters, session_required_policies has
expanded the set of types it accepts from `str` to `list[str]`.

* Make authorization_parameters more type-safe

When parsing authorization_parameters, avoid previous assignments
which assumed that values had the correct types. Instead, always check
values against the expected type(s) and convert them to None with a
warning log if they are malformed.

In service of this improvement in strictness, the class has been
refactored to have a more verbose but highly unambiguous structure.

* Make `_is_list_of_strs` a TypeGuard

This lets mypy understand its meaning vis-a-vis types.