Support session params in oauth2_get_authorize_url (#773)
Add explicit support for a set of known parameters to the
oauth2_get_authorize_url method. These are all some flavor of string,
uuid, or list thereof, and are therefore converted to comma separated
lists by the SDK.

The parameter descriptions are kept very brief for now, to avoid any
misleading statements about what these parameters do.

A test case is established for these options, building a subset of the
full parameter matrix (which produced hundreds of unnecessary tests).