Implement `AuthClient.create_project` (#772)
* Implement `AuthClient.create_project`

Also expand utils.safe_strseq_iter to allow for UUIDs as inputs. This
simplifies the implementation of the `admin_ids` transform.

* Add explicit tests for safe_strseq_iter

Also make it handle `bytes` by doing an implicit decode. Although this
is not part of our advertised interface anywhere, handling bytes means
that we will have a more graceful behavior in APIs where this is used
if a user passes a bytestring at runtime.

* Remove bytes handling from safe_strseq_iter

This is nowhere part of documented SDK contracts. Rather than having
"hidden" handling for bytes, refuse to code around the possibility.
Users are responsible for *not* passing bytes.