Add tests for paginator signature matching (#779)
These tests enforce that decorated methods of the target client
classes satisfy the paginator interfaces for their decorators.

For example, if a method is decorated with the MarkerPaginator, then
it must take a named argument `marker` for the paginator to use.

It's acceptable for the argument to be keyword-only or for it to be
positional-or-keyword. But it's not acceptable for it to be, e.g.,
positional-only.