Add search index lifecycle methods: create_index, delete_index, reopen_index (#785)
* Add methods for index {create,delete,reopen}

These are a trio of Search API methods. Here they are introduced
without test data, tests, or anything beyond their basic call
structure and docs.

Because the docs reference fixture data which does not yet exist, the
docs are not fully valid yet.

This change includes a changelog for the new methods.

* Add test data for Search index lifecycle APIs

* Add simple tests for Search index lifecycle APIs

* Add failure tests to new Search methods

`{create,delete,reopen}_index` now each have an interesting failure
mode recorded in the test data and each one is tested to ensure that
the errors parse correctly.