Add BaseAuthClient to separate methods
Introduce a new base class for the Auth client hierarchy in order to
split the developer API methods from the previously existing "base"
methods. The important exception is `get_identities`, which supports
client credentials authorization and therefore needs to stay on the
base.

The changelog explains this change without drawing undue attention to
the ways in which it could, in theory, be compatibility breaking for
users relying on the inheritance relationship (e.g. via `isinstance`
or type annotations).

No new tests are added and no tests are broken by this change.