Add Auth Manage Projects example scripts (#777)
* Add Auth Project management examples to docs

* In Update Project example, do not set admin_ids

Due to a bug, admin_ids or group_admin_ids must be supplied when
performing an update. As a result, the prior version of this script
sets the current user as an admin and this necessitates warning text.

Remove that setting to avoid having a warning which is based around a
bug. This means that the example cannot be considered ready to release
until the bug is resolved.

* Remove "update" from manage-projects example

This is blocked on a bug in the way that updates are handled vis-a-vis
admin_ids. Rather than holding back the whole doc, simply drop the
`update` functionality. It can be restored in the future.

* Remove args from list_projects example functions