Rename a variable to avoid spurious pylint failure
pylint flagged `graph` (used earlier in a function) as a redefinition
of a name from outer scope. This happens because the block under the
`if __name__ == "__main__"` check is technically in module scope.
Therefore, at analysis-time, there is a module-scoped variable named
`graph` and we "need" to avoid redefinition.

This could be tucked into a `def _main()` to "hide" the name from
module scope, but a rename is equally simple to dodge the
overly-aggressive pylint check.