Add more tests and rewrite cycle detection
In the experimental scope graph component, the cycle detection code
was not correct. The simplicity of the approach used did not properly
account for a number of scenarios like `foo[bar[baz] baz]`, which
would be incorrectly detected as a cycle. These errors were masked
by a secondary flaw, an incorrect call to `set.clear()` which wiped
out important information.

Rather than continuing to pursue this approach, for a first draft of
this component simply use a well-understood recursive cycle detection
algorithm which operates linearly in the size of the graph.

The implementation can and should be unrolled to be iterative in the
future.