Revert cycle detection to a simpler implementation
Our goal is still correctness, but the use of Tarjan's strongly
connected components algorithm makes the ScopeGraph class
significantly more complex in pursuit of efficiency.
Worse, the prior implementation was recursive, which carries
significant costs in python.

Revert to a simpler iterative cycle detector which is adjusted
to be certain to be correct but somewhat inefficient.