Refactor GARE validators to protocol system
Declare a well-typed protocol for validators, which defines the
`__call__` method to take a name (for the field) and a value to
validate.

Each field validation is now done with an explicit call to a
validator, and without any derivation of validators, types, or
meta-programming.

The tow remaining items which needed special handling were:
- checking for non-null values being passed for at least one argument
- loading a data object from a dict filtering out extra params

In these two cases, the desired arg names are found by examining
`__init__` and excluding `self` and `extra`, each of which has special
meaning.

In general, the paradigm shift offered here is to be more direct and
explicit, even if the requirement is that we are verbose in the
process. Partly, this was based on the realization that many of the
meta-programming techniques being used were not resulting in the
desired level of brevity and clarity.