Remove runtime use of signature inspection
Relegate the strategy of looking at `__init__` signatures to the
testsuite. Rather than doing this at runtime, reintroduce
`SUPPORTED_FIELDS` as a property of GARE-related classes, but as a set
of strings.

The testsuite now checks the set against `__init__` for the relevant
class.
The result is more verbose but simpler.