Add a script to get Requires-Python metadata
This does a temporary sdist build and extracts the result in order to
get the (singular) '>=' lower bound.

If the expectations of the script are violated, an error is raised.

The approach is to be "slow but robust" here. We could simply parse
`setup.cfg`, but that makes the toolchain sensitive to changes in
packaging tooling. The PKG-INFO data should be reliable in all cases.