Put the 'get_python_requires' venv cache in .tox/
Allow passing a cache directory explicitly to
`get_python_requires.py`. Subsequently, the wrapping script,
`ensure_min_python_is_tested.py` can insist on an env var which
provides the cache dir path.
Finally, `tox.ini` gets a final tweak to set this value via setenv to
the `envdir` (the virtualenv dir which also contains the tmp and log
dirs).

The choice of an env var over parsing an argument is merely that it is
marginally simpler to read an env var and fail on its absence than it
is to do the same with an argument. (At the very least, *any* argument
parsing opens the question of whether or not argparse should be used.)