Suppress pylint docparams warnings in client.py
In many other cases missing parameters from the docstring represents a
problem we need to fix. In this case, however, we have everything
documented in one place, and then other methods refer to it ("see
`request` params"). We could expand out the docs in all of them, but
it makes the experience reading the docs worse, rather than better.
In this unusual case, suppress the warning.

For our `resource_server` classproperty, a slightly different warning
is appropriate to suppress. pylint cannot understand the decorator as
a descriptor, so it expects documentation of an implicit parameter
(which, as with the other case, would be detrimental to the docs).