Expand 'typing' reqs to all python versions
We were building our pinned requirements data for typing only
under the python version used for `mypy` linting in CI: Python 3.11.

This makes mypy-linting of code non-functional on lower interpreter
versions. While some SDK developers are using pyenv or a similar
solution to install multiple pythons in their home dirs, so we didn't
notice the constraint, this is not meant to be a requirement for all
contributors.

To make `tox r -e mypy` work on all pythons, the `tox` target for
"freezedeps" has been modified to generate the extra requirements,
and the freezedeps job has been run.

The fix was verified with

    tox r -e mypy --discover $(which python3.7)
    tox r -e mypy --discover $(which python3.9)

It is possible that this will cause some drift and issues, especially
around our lower-bound for python version (e.g. latest `mypy` drops
support for a version we use), but at present such issues are
speculative. We can deal with additional tightening of bounds and
adjustments to CI to run some of these versions in the future.