Introduce 'TransferTimer' and 'create_timer'
This is a new payload type and helper for submitting Transfer Timers.

In addition to introducing the new types, this change officially
deprecates all use of `TimerJob.from_transfer_data`.
Existing tests which use this form are simply marked with
`pytest.warns` context managers to keep them functional through this
changeset.