Error if TimerClient is given wrong payload types
In order to catch the likely confusion across the punnet square of

+--------------+----------+---------------+
|              | TimerJob | TransferTimer |
+--------------+----------+---------------+
| create_job   |   good   |   bad         |
+--------------+----------+---------------+
| create_timer |   bad    |   good        |
+--------------+----------+---------------+

the "bad" cases now emit `GlobusSDKUsageError`s.

For TimerJob x create_timer, the user is instructed to create a
TransferTimer instead.
For TransferTimer x create_job, the user is instructed to use
create_timer instead.

In each case, we are guiding users towards newer usage.