Automatically strip MISSING from payloads
This change updates payload serialization in two significant ways:

- MISSING is automatically removed from payloads, and this is applied
  recursively. This means that it can be included anywhere in a
  (well-formed) body and the SDK will find and remove it. Custom types
  inside of the payload could interfere, but native dict/list
  structures and PayloadWrapper objects are fully supported

- PayloadWrapper types are recursively converted to dicts, meaning
  that it is now valid to construct a helper type based on
  PayloadWrapper which is nested

As a result of this change, simplify the GroupPolicies code to always
assign the ha timeout (with a default value of MISSING) and replace
the unit test for this behavior with a functional test.

The new recursive conversion was benchmarked against a version which
was micro-optimized to avoid deep-copying of dict and list data, and
found to be nearly 2x *faster* than the optimized version on a typical
payload, by virtue of the fact that it did not make unnecessary method
calls.