Add support for `MISSING` in params and headers
This is an internal feature of the client class.
Entirely independently from the payload filtering process (which needs
to handle the complex payload types which may be used), a simple
helper for `dict|None` data is used to strip out MISSING from any
query params and headers before sending.

Testing also confirms the behavior of MISSING on simple JSON and
Form-POST data.