Make 'experimental' more lazy-import friendly
Modules under 'experimental' are now tested -- as part of the
lazy-import dedicated tests -- to confirm that they do not have any
import-time dependency on 'requests'. This helps ensure that these
modules are safe to use in an application like globus-cli which is
extremely latency-sensitive for slow imports, without accidentally
incurring the large import-time costs of the `requests`, `urllib3`,
etc dependency chain.

In order to pass the test, one import is moved to be
TYPE_CHECKING-deferred in the GARE modules.