Automatically convert UUID to str in requests
The `RequestEncoder` classes define how we encode payloads and
otherwise "prepare" a request. They are therefore a better location
for any transforms we want to perform between some client call being
made (i.e. `post()`, `get()`, ...) and the call to `requests`.

Therefore, the change has two major pieces:
- Move all transformation of request data from `BaseClient` and
  `PayloadWrapper`. This covers removal of MISSING from all three
  major locations for request data.
- Add the capability to auto-stringify UUIDs to the encoder passes.
  (With the first part done, this is near-trivial.)

Note that the names used in the encoders for these data are
intentionally consistent with `requests.Request` fields as much as
possible. So we have internal hooks for `_prepare_params`,
`_prepare_headers`, and `_prepare_data` (that last one being the
payload/body).

Tests which were written for the PayloadWrapper have been replaced
with tests for the encoders.

`utils.filter_missing` is still defined separately because it was used
in some tests to confirm the removal of `MISSING` from data -- it is
therefore demonstrably useful independent from the other changes.
That is why it is defined separately and used as part of the
transformations, rather than being inlined there. (Inlining would make
things marginally faster, but carries tradeoffs.)