Improve the way that enum members get stringified
The GroupsClient and related helpers have a number of enums of strings
which are accepted as inputs. Previously, rendering these enums to
their `.value` attributes (their values) was done via explicit calls
in all of the locations where they are used.

Under this change, the serialization of any `Enum` member as `x.value`
is codified as one of the characteristics of the primitive-value
processing done in the transport layer's encoders, alongside UUID
stringification.

The end result is a minor but useful simplification to any code which
handles these enums, as the enum value can be passed to the underlying
transport without a prior transformation. A minor helper function is
eliminated (in favor of the addition to the encoders).

One test, which evaluated the correct stringification of an enum, has
been updated to use `responses` to check the sent request, rather than
a payload object prior to sending.