TransferClient relies on UUID encoding transport
Instead of explicitly stringifying values in TransferClient, allow
UUIDs to be passed to the transport in all cases.

Additionally, make use of `utils.commajoin` as a replacement for
custom and inlined behaviors which are identical in meaning.

The goal of this change is to be a small net-negative diff -- not only
in lines, but also in some cases in complexity on the same number of
lines.