Auth app clients trust more to the transport
Update NativeAppAuthClient and ConfidentialAppAuthClient to trust more
of the work of encoding a payload to the transport layer.
Specifically, trust it more for handling UUIDs and MISSING.

This is a reduction in LOC and complexity and a good demonstration of
the value gains from a more capable custom encoder underneath the
client layer.