Service auth client trusts more to the transport
This is an analogous change to #896, making the service client rely
more heavily on the fact that the transport will strip MISSING and
translate UUIDs.

There are a number of cases here which are related to but not covered
by the current toolchain for handling MISSING. In particular, checks
for an "empty" subdocument (see `links` in client create/update) and
handling of "commajoin" on a list which may be MISSING are both places
where utility improvements may allow even further simplifications to
the client code. However, given that there is no change possible
today, no further adjustment is made to those parts of the client
behavior.