Enhance docs for 'client_type'
1. Reorder the keyword-only parameters for client creation to put
   `client_type` adjacent to `public_client`

This has no semantic impact. It's done only to make the parameter
order in code match the order in doc.

2. Reorder parameter docs to make sure `client_type` is adjacent to
   `public_client`

These are mutex and one is required, which makes me want them near
each other and at the top of the parameter list.

3. Convert doc of `client_type` values to a list-table under a
   dropdown

I find this to be a good balance of information, vertical density for
scrolling, and easy authorship. `list-table` is pretty simple and
requires minimal clever ASCII-art arrangements (vs many other table
systems for markup languages).