Add a changelog for recent Auth client changes
This is a post-facto changelog for #884, with all of the methods
enumerated with a short description.

While writing the changelog, it became clear that users could get
confused about "client" vs "Client" vs `AuthClient`, etc, so an
introductory note is included in this changelog entry -- the intent is
for it to be included in the permanent changelog as-is, under the
`Added` items for this release.