Add `delete_protected` for mapped collections
This is a new field in the MappedCollectionDocument definition.

As a bit of new testing art to try to validate the change, the unit
tests for the GCS helpers now include some signature-inspection tests
+ a test for optional bool settings. These are rather simplistic
tests which might not offer a ton of value, but they're also very
quick to execute.