Remove unnecessary internal dict-conversions
The SDK had a number of locations in which it would convert a
PayloadWrapper class into a dict so that it could be used as an
internal component in a request payload.

Now that the request encoders handle this automatically, it's not
necesary to eagerly convert an input object into a dict in order for
it to be serialized correctly.
Defer the dict conversions until request sending by removing the
unnecessary eager calls to `dict(obj)`.

Tests which check on the results of encoding are refined to use an
encoder to build a dummy `Request` object, but are otherwise
unchanged.