Convert Auth components to new autodoc typehints
- remove `:type ...:` and `:rtype ...:` documentation
- set `autodoc_typehints_description_target="documented_params"` to
  improve client class initializer documentation
- reorder unions of `MissingType|Literal[...]` in order to avoid a
  sphinx bug which appears to incorrectly render a Literal union in
  which the Literal is the left-hand side
- alter ScopeCollectionType to allow autodoc inspection

Regarding ScopeCollectionType, the basic change is that we now allow
`globus_sdk._types` to import `MutableScope` at runtime and move the
deferred type-checking imports into `globus_sdk.scopes`. As a result,
the runtime-visible value contains the proper `MutableScope` object,
not a string forward reference (which sphinx doesn't know how to
resolve). This slight rearrangement basically makes sphinx work.

Additionally, since we were changing the type, remove two elements of
the union which are redundant. Most likely, this had previously been
written with `list`, which is invariant, and was not correctly fixed
when it was widened to `Iterable`.