Update Groups to autodoc type hints
- remove all uses of `:type:` docs
- add `ArrayResponse` (which was missing) so that doc linking works
  for rtype and docs are more complete
- in the case of `GroupPolicies`, special handling produces UnionType
  formatted strings, rather than allowing autodoc logic to run

For whatever reason, autodoc does not correctly process a union when
it gets the union type via the signature of `__init__`, and can't
resolve the type alias there in the same way that it does for methods.
To work around this for now, the docstring of the class is updated to
be a non-interpolated format-string, which is then converted at
runtime using `.format()` via a decorator.

This keeps the string definition of these unions local to the place
where they are defined and used.