Conver paging to autodoc type hints + fix bug
Paging classes documented the type of the argument tuple correctly,
but then provided an annotation for it as `list[Any]`.

Fix this by making it properly a tuple (which is what it was always
intended to be).
This constitutes a minor change at runtime from a list which is never
modified to a tuple, but most importantly fixes the incorrect type
annotations.