Improve how paginator contracts are tested
Paginators now provide a class attribute (`_REQUIRES_METHOD_KWARGS`)
which is a tuple of the keyword args which are expected on the
paginated methods. The contract test for paginated methods can
therefore iterate over all paginated client methods and check this
tuple against the kwargs.

This eliminates explicit mapping from paginator types to these
sets of argument names in the testsuite.