Define 'extdoclink' role and use it for Search
The end-goal here was to add several link references to the Search
docs for the `advanced` query string syntax, which is mentioned in
these docs but not fully explained.

In a few docstrings, we're adding links to the Search API docs.

In support of this, define an `extdoclink` role, in the custom sphinx
extension, which can be used as a simplified version of the `extdoclink`
directive:
- extract URL deduction into a helper for use in both
- do some very simplistic parsing to make the role easy to implement
  and read
- do not support all of the features of the extdoclink directive --
  e.g. you cannot customize `service` at present
- enable the role via `app.add_role` in sphinx `setup`