Allow passing `bytes` for payload bodies
This change is only applied to the "raw" HTTP methods. Higher level
methods, belonging to per-service clients, still operate in terms of
strings and do not complicate their signatures with bytes.

This requires various type signatures to be updated in the transport
layer. Most notably, a check for `isinstance(X, str)` has been
expanded to `isinstance(X, (bytes, str))`, meaning that `bytes` are
now treated similarly to `str` when choosing how data should be
encoded as JSON, urlencoded, or passed as raw text.