Replace automodule directives with autoclass
automodule doesn't use `current_module` and therefore shows the full
source-path for the classes it documents. We export almost all of
our user-facing names in our top-level `__init__.py` and therefore
want to document with the `globus_sdk` module name. The only solution
which seems to work cleanly is to avoid automodule and use autoclass
instead.

Advantageously, this provides two additional benefits:
- it reveals that some of the GCS-related response types were missing
  from our exports
- it allows us to group the contents of the documented modules under
  headings to better explain the content to users

The change in our exports gets a changelog note.