Add 'query_params' to token introspect
This is a passthrough to allow arbitrary parameter passthrough. It is
being added based on a conversation with the Auth team indicating that
introspect may support query parameters in the future, and that this
passthrough will be useful for development of such parameters.

Importantly, this adds two tests + test data for introspect, to ensure
that passing no values for query params does not send query params,
and that params can be passed successfully.