Update JSON Token Storage to use TypedDict
- Apply TypedDict to both the mainline and experimental variants to
  declare the expected/required fields for valid data more explicitly
- Expand the error checking in both of these variants to match the
  data loaded against the declared TypedDict invariants

This changeset was blocked on the removal of py3.7 and is noted in
existing comments as requiring 3.8+ . 3.8 adds TypedDict, but it is
not clear why `typing_extensions` was not used to provide it. Perhaps
the original comment predates our inclusion of `typing_extensions` in
globus-sdk dependencies.

The changeset could either be considered an addition, an enhancement,
or a bugfix. It is therefore qualified under the most generic
changelog category of "change".