Remove 'config' from new sqlite tokenstorage
Support for the 'config_storage' table, and associated methods, has
been removed after deliberation about the modeling of tokenstorage
adapters and their interchangeability.

The issue with config storage is that none of the other adapters
provide and support this interface. Therefore, what appears to be
generic and pluggable is, in fact, a locked-in decision for certain
applications (e.g., `globus-cli`) and will not allow us to provide
swappable token storage for those applications.

Such extension points can easily be patched into place for migration
purposes in the few applications which need them, conditional on the
presence/absence of the `config_storage` table. Developing migration
workflows is left as a future problem.