Reject ':memory: in sqlite tokenstorage v2
Use of a ':memory:' database indicates an opportunity to use the
memory tokenstorage. We now reject ':memory:' with an appropriate and
detailed usage error.

At the same time, various uses of 'dbname' rather than 'filename' have
been conformed to 'filename'. This further reinforces that the
tokenstorage object is meant to refer to some known file, not a pure
memory DB.

Tests which relied on the memory database have been updated to use a
tempfile. One test which needed significant changes was completely
refactored at this time, to provide better and clearer testing.