Move 'ensure dir' capability to FileTokenStorage
Rather than this being a specialized characteristic of the
JSONTokenStorage, this is now part of the base class functionality. As
a result, both the JSON and SQLiteTokenStorage inherit this behavior
(with a small change to the SQLiteTokenStorage initializer to call
super() at the right time).

This allows both of these storage adapters to inherit this beneficial
behavior.