Use tmp_path rather than tempfile in tests
For testing purposes, rather than manually managing temporary files
ourselves, rely on the standard `tmp_path` pytest fixture to do this.
It appears to be more robust in some circumstances than the
on-yield-teardown behavior, and reduces the amount of code which we
are maintaining.

Most changes here are name changes and conversions from strings to
pathlib.Path objects (as is convenien with the new fixture).
Slightly more changes are applied to the GCP object tests, since those
make more elaborate use of the temporary directory they manage.