Move task_list tests to new paradigm
The existing tests for `TransferClient.task_list` are old, and
use testing tools which predate `globus_sdk._testing`. We can update
this to our newer paradigm and move the tests to a dedicated module
without much effort. The test payload content has only been changed a
small amount, and the tests are preserved with near identical behavior
to their earlier incarnation.