Add ConnectorTable.extend() method
This method allows ConnectorTable consumers to extend the class with
an additional connector (or, if they chain calls, multiple connectors).
The new connector is written into the new subclass as a part of its
`_connectors` tuple and as an attribute.

The base ConnectorTable class is left unmodified, meaning that callers
who extend the table cannot conflict.

A usage example is provided in the docstring for `extend()` to
instruct users on proper usage.