Add UUID checking for add_app_data_access_scope
In order to help users catch avoidable errors during login flows, this
adds a validation check which confirms that the UUID given to
`add_app_data_access_scope` is a valid UUID str. It also checks type
-- guarding against `None` producing strings like
`.../None/data_access` and similar type errors.

The check is defined in the utils module for potential reuse, but no
extra usage sites are added in this changeset.